requirements.txt
asgiref==3.7.2
attrs==25.1.0
bleach==6.0.0
Django==4.2.1
django-appconf==1.0.5
django-compressor==4.4
django-debug-toolbar==4.1.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-mail-templated==2.6.5
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.18
django-summernote==0.8.20.0
django-taggit==4.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
importlib_metadata==8.6.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
Pillow==10.0.0
PyJWT==2.9.0
PyYAML==6.0.2
rcssmin==1.1.1
referencing==0.36.2
rjsmin==1.2.1
rpds-py==0.23.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
webencodings==0.5.1
zipp==3.21.0
